move all sectors to subfolders
