+++ generatorcore/electricity2018/__init__.py
@@ -1,7 +1,7 @@
 from dataclasses import dataclass, field
-from .inputs import Inputs
-from .utils import div
-from . import transport2018
+from ..inputs import Inputs
+from ..utils import div
+from .. import transport2018
 
 # Definition der relevanten Spaltennamen für den Sektor E
