+++ generatorcore/residences2018/__init__.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass, field
-from .inputs import Inputs
-from .utils import div
+from ..inputs import Inputs
+from ..utils import div
 
 
 @dataclass
